server/requirements.txt
fastapi==0.136.0
groq==1.2.0
langgraph==1.1.9
pydantic==2.13.3
python-dotenv==1.2.2
structlog==25.5.0
tenacity==9.1.4
uvicorn
supabase
